Rename AtomicCounter to TaskCounter, and remove all the public member functions
Create a copy called FullAtomicCounter that *does* have all the atomic counter
related functions.

This allows us to do optimizations for the 99% of cases where counters are just
used for task waiting. However, this makes TaskCounter polymorphic, forcing a
vtable lookup. That said, the large majority of TaskCounter instances should be
just the base class, so the branch predictor should be right most of the time.